chore: 라이브러리 업데이트, 스니펫 최신화
