+++ .env
@@ -1,1 +1,1 @@
-<!-- # 공통 환경 변수 -->
+# 공통 환경 변수
