--- .storybook/preview.jsx
@@ -1,22 +1,0 @@
-import '../src/index.css';
-import { BrowserRouter } from 'react-router-dom';
-
-export const parameters = {
-  actions: { argTypesRegex: '^on[A-Z].*' },
-  controls: {
-    matchers: {
-      color: /(background|color)$/i,
-      date: /Date$/,
-    },
-  },
-};
-
-export const decorators = [
-  (Story) => (
-    <>
-      <BrowserRouter>
-        <Story />
-      </BrowserRouter>
-    </>
-  ),
-];
