+++ package.json
@@ -18,45 +18,42 @@
     "cli": "jordy-cli",
     "test": "vitest run",
     "test:watch": "vitest watch",
-    "sb": "start-storybook -p 6006",
-    "build:sb": "build-storybook",
+    "sb": "storybook dev -p 6006",
+    "build:sb": "storybook build",
     "serve:sb": "npx http-server ./storybook-static -o",
     "cy": "cypress open",
     "cy:run": "cypress run"
   },
   "dependencies": {
-    "@reduxjs/toolkit": "^1.9.0",
-    "axios": "^1.1.3",
-    "jordy": "^0.17.1",
+    "@reduxjs/toolkit": "^1.9.5",
+    "axios": "^1.4.0",
+    "jordy": "^0.21.2",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
-    "react-redux": "^8.0.5",
-    "react-router-dom": "^6.4.3",
-    "styled-components": "^5.3.6",
-    "tslib": "^2.4.1"
+    "react-redux": "^8.1.1",
+    "react-router-dom": "^6.14.2",
+    "styled-components": "^6.0.5"
   },
   "devDependencies": {
     "@babel/core": "^7.20.2",
-    "@storybook/addon-actions": "^6.5.13",
-    "@storybook/addon-essentials": "^6.5.13",
-    "@storybook/addon-links": "^6.5.13",
-    "@storybook/builder-vite": "^0.2.5",
-    "@storybook/react": "^6.5.13",
-    "@storybook/testing-library": "^0.0.13",
-    "@testing-library/jest-dom": "^5.16.5",
-    "@testing-library/react": "^13.4.0",
+    "@storybook/addon-actions": "^7.1.0",
+    "@storybook/addon-essentials": "^7.1.0",
+    "@storybook/addon-links": "^7.1.0",
+    "@storybook/addon-mdx-gfm": "^7.1.0",
+    "@storybook/react": "^7.1.0",
+    "@storybook/react-vite": "^7.1.0",
+    "@storybook/testing-library": "^0.2.0",
+    "@testing-library/jest-dom": "^5.17.0",
+    "@testing-library/react": "^14.0.0",
     "@types/node": "^18.11.9",
-    "@types/react": "^18.0.25",
-    "@types/react-dom": "^18.0.9",
-    "@types/react-redux": "^7.1.24",
-    "@types/react-router-dom": "^5.3.3",
-    "@types/styled-components": "^5.1.26",
+    "@types/react": "^18.2.15",
+    "@types/react-dom": "^18.2.7",
     "@typescript-eslint/eslint-plugin": "^5.43.0",
     "@typescript-eslint/parser": "^5.43.0",
-    "@vitejs/plugin-legacy": "^2.3.1",
-    "@vitejs/plugin-react": "^2.2.0",
+    "@vitejs/plugin-legacy": "^4.1.1",
+    "@vitejs/plugin-react": "^4.0.3",
     "babel-loader": "^9.1.0",
-    "cypress": "^11.1.0",
+    "cypress": "^12.17.2",
     "env-cmd": "^10.1.0",
     "eslint": "^8.27.0",
     "eslint-config-prettier": "^8.5.0",
@@ -65,15 +62,17 @@
     "eslint-plugin-jsx-a11y": "^6.6.1",
     "eslint-plugin-react": "^7.31.10",
     "eslint-plugin-react-hooks": "^4.6.0",
-    "eslint-plugin-storybook": "^0.6.7",
-    "jordy-cli": "^0.5.0",
-    "jsdom": "^20.0.2",
+    "eslint-plugin-storybook": "^0.6.13",
+    "jordy-cli": "^0.7.0",
+    "jsdom": "^22.1.0",
     "prettier": "^2.7.1",
+    "storybook": "^7.1.0",
     "terser": "^5.15.1",
-    "typescript": "^4.9.3",
-    "vite": "^3.2.4",
-    "vite-tsconfig-paths": "^3.5.2",
-    "vitest": "^0.25.2"
+    "tslib": "^2.4.1",
+    "typescript": "^5.1.6",
+    "vite": "^4.4.7",
+    "vite-tsconfig-paths": "^4.2.0",
+    "vitest": "^0.33.0"
   },
   "author": "",
   "license": "ISC",
